requirements.txt
# --- START OF FILE requirements.txt ---
flask
flask-cors
gunicorn
google-genai # <--- ADD THIS
# --- END OF FILE requirements.txt ---